+++ source/B2B.Transactions.Api/Configuration/Middleware/Correlation/SessionContext.cs
@@ -14,7 +14,7 @@
 
 using System;
 
-namespace B2B.Transactions.Api.Middleware.Correlation
+namespace B2B.Transactions.Api.Configuration.Middleware.Correlation
 {
     public sealed class SessionContext : ISessionContext
     {
