+++ source/B2B.Transactions.Api/Configuration/Middleware/ServiceBus/ServiceBusMessageMetadataMiddleWare.cs
@@ -21,7 +21,7 @@ using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Middleware;
 using Microsoft.Extensions.Logging;
 
-namespace B2B.Transactions.Api.Middleware.ServiceBus
+namespace B2B.Transactions.Api.Configuration.Middleware.ServiceBus
 {
     public class ServiceBusMessageMetadataMiddleWare : IFunctionsWorkerMiddleware
     {
