+++ source/B2B.Transactions.Api/Program.cs
@@ -15,10 +15,10 @@
 using System;
 using System.IdentityModel.Tokens.Jwt;
 using System.Threading.Tasks;
-using B2B.Transactions.Api.Middleware.Authentication.Bearer;
-using B2B.Transactions.Api.Middleware.Authentication.MarketActors;
-using B2B.Transactions.Api.Middleware.Correlation;
-using B2B.Transactions.Api.Middleware.ServiceBus;
+using B2B.Transactions.Api.Configuration.Middleware.Authentication.Bearer;
+using B2B.Transactions.Api.Configuration.Middleware.Authentication.MarketActors;
+using B2B.Transactions.Api.Configuration.Middleware.Correlation;
+using B2B.Transactions.Api.Configuration.Middleware.ServiceBus;
 using B2B.Transactions.Infrastructure.Configuration;
 using B2B.Transactions.Infrastructure.OutgoingMessages;
 using Energinet.DataHub.Core.Logging.RequestResponseMiddleware;
