+++ source/B2B.Transactions.IntegrationTests/TestBase.cs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 using System;
-using B2B.Transactions.Api.Middleware.Correlation;
+using B2B.Transactions.Api.Configuration.Middleware.Correlation;
 using B2B.Transactions.Infrastructure.Configuration;
 using B2B.Transactions.IntegrationTests.Fixtures;
 using B2B.Transactions.IntegrationTests.TestDoubles;
